Moving the provider specific code into its own class
